spike adaption

+++ noise_analysis/create_network.py
@@ -47,7 +47,7 @@ def create_arrays(N, resting_membrane, total_time, max_time, data, N_x):
     trace = np.zeros((total_time, N))
     trace[0] = 1
 
-    spikes = np.zeros((total_time, N))
+    spikes = np.zeros((total_time, N), dtype="int64")
     spikes[:, :N_x] = data
 
     spike_times = np.random.randint(low=max_time, high=max_time**2, size=N)
